pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "ipykernel>=7.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.26.0",
    "pypdf>=6.8.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "tiktoken>=0.12.0",
]
